move PersonalDetails and add profile id to saved data
