+++ packages/chealt.com/public/App/state.js
@@ -1,0 +1,11 @@
+import { createContext } from 'preact';
+
+import { state as profiles } from '../Profiles/signals';
+
+const createAppState = () => ({
+  profiles
+});
+
+const AppState = createContext();
+
+export { createAppState, AppState };
