--- packages/chealt.com/public/App/utils.js
@@ -1,7 +1,0 @@
-import { state as profiles } from '../Profiles/signals';
-
-const createAppState = () => ({
-  profiles
-});
-
-export { createAppState };
